requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
tf-keras
faiss-cpu
langchain_chroma
langchain-openai
langchain_groq
langchain_core
fastapi
uvicorn
langserve